+++ crates/cubecl-linalg/src/matmul/components/global/multi_buffer/double_buffering/base.rs
