+++ crates/cubecl-linalg/src/matmul/components/global/multi_buffer/loading/async/maximize_slice_length.rs
@@ -1,6 +1,6 @@
 use crate::matmul::components::{
     global::{
-        single_stage::AsyncBufferLoadingStrategy,
+        single_buffer::AsyncBufferLoadingStrategy,
         tensor_view::{TensorReader, Window},
         CopyMechanism, GlobalConfig, LoadingValidation,
     },
