+++ crates/cubecl-linalg/src/matmul/components/global/multi_buffer/specialized/config.rs
