+++ crates/cubecl-linalg/src/matmul/components/global/multi_buffer/specialized/matmul.rs
@@ -1,6 +1,6 @@
 use crate::matmul::components::global;
-use crate::matmul::components::global::multi_stage::double_buffering::BufferId;
-use crate::matmul::components::global::multi_stage::{
+use crate::matmul::components::global::multi_buffer::double_buffering::BufferId;
+use crate::matmul::components::global::multi_buffer::{
     BufferLoader, SyncBufferLoader, SyncBufferLoadingStrategy,
 };
 use crate::matmul::components::global::output_loader::Unloader;
