+++ crates/cubecl-linalg/src/matmul/components/global/single_buffer/config.rs
