+++ crates/cubecl-linalg/src/matmul/components/global/single_buffer/loading/async/maximize_slice_length.rs
