+++ crates/cubecl-linalg/src/matmul/components/global/single_buffer/loading/async/maximize_unit_count.rs
