+++ crates/cubecl-linalg/src/matmul/components/global/single_buffer/simple/mod.rs
